[refactor] #20 SearchResultView 의존성 주입
